Time: 21 ms (64.31%), Space: 17.6 MB (21%) - LeetHub

+++ 0018-4sum/0018-4sum.cpp
@@ -16,7 +16,7 @@ public:
                 int c = b+1;
                 int d = n-1;
                 while(c < d){
-                    signed long long val = nums[a] + nums[b] + nums[c] + nums[d];
+                    long long val = (long long)nums[a] + nums[b] + nums[c] + nums[d];
                     if(val == target){
                         vector<int> temp = {nums[a] , nums[b] , nums[c] , nums[d]};
                         answer.emplace_back(temp);
